fix(system-tests): extract common commands
Extract common child commands between integration and system tests.
Small tweaks to get integration tests up to date.
