+++ system-tests/services/test-external-volumes.js
@@ -1,5 +1,5 @@
 require("../_support");
-require("../_support/utils/ServicesUtil");
+require("../../tests/_support/utils/ServicesUtil");
 const { Timeouts } = require("../_support/constants");
 
 describe("Services", function() {
