+++ system-tests/services/test-pods.js
@@ -1,4 +1,4 @@
-require("../_support/utils/ServicesUtil");
+require("../../tests/_support/utils/ServicesUtil");
 const { Timeouts } = require("../_support/constants");
 
 describe("Services", function() {
