+++ tests/pages/nodes/node/tasks/NodesTaskDetailPage-cy.js
@@ -9,8 +9,8 @@ describe("Nodes Task Detail Page", function() {
   context("Navigate to node task detail page", function() {
     it("navigates to node task detail page", function() {
       cy.visitUrl({ url: "/nodes", identify: true, fakeAnalytics: true });
-      cy.get("tr a").eq(0).click();
-      cy.get("tr a").eq(1).click();
+      cy.get("a.table-cell-link-primary").eq(0).click({ force: true });
+      cy.get("a.table-cell-link-secondary").eq(0).click();
       cy.hash().should("match", /nodes\/[a-zA-Z0-9-]+\/tasks\/[a-zA-Z0-9-]+/);
 
       cy.get("h1.configuration-map-heading").contains("Configuration");
@@ -18,8 +18,8 @@ describe("Nodes Task Detail Page", function() {
 
     it("loads page with data after hard reload", function() {
       cy.visitUrl({ url: "/nodes", identify: true, fakeAnalytics: true });
-      cy.get("tr a").eq(0).click();
-      cy.get("tr a").eq(1).click();
+      cy.get("a.table-cell-link-primary").eq(0).click({ force: true });
+      cy.get("a.table-cell-link-secondary").eq(0).click();
       cy.hash().should("match", /nodes\/[a-zA-Z0-9-]+\/tasks\/[a-zA-Z0-9-]+/);
 
       cy.reload(true);
