+++ tests/pages/services/ServiceExternalVolumes-cy.js
@@ -1,4 +1,4 @@
-require("../../_support/utils/ServicesUtil");
+require("../../../tests/_support/utils/ServicesUtil");
 
 describe("Services", function() {
   /**
